Many fixes
 - add eslint and fix up issues
 - rename Result.state to .status and drop the 'ed' from the statuses
 - add a new useProperties() hook for fetching state as a subset of the
   properties of a given item, driven via the 'notify' event

useProperties() has a couple of issues, the main one being that it
contains an effect which runs every time, for data dependency reasons.

One of those is related to the exact event handling we use, and could be
fixed via useEffectEvent() which is coming in some future React release.

The other of those problems is related to the fact that the keys list
doesn't identity-compare to itself between iterations and therefore
would always cause a reload if used as a dependency.